refactor: 안쓰는 dto 삭제
