+++ server/src/main/java/moment/moment/dto/response/MyMomentEchoResponse.java
@@ -1,7 +1,6 @@
 package moment.moment.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import moment.comment.domain.Echo;
 import moment.comment.dto.tobe.EchoDetail;
 
 @Schema(description = "내 모멘트의 코멘트의 에코 조회 응답")
@@ -15,10 +14,6 @@ public record MyMomentEchoResponse(
         @Schema(description = "조회된 Echo 작성자")
         String userName
 ) {
-
-    public static MyMomentEchoResponse from(Echo echo) {
-        return new MyMomentEchoResponse(echo.getId(), echo.getEchoType(), echo.getUser().getNickname());
-    }
 
     public static MyMomentEchoResponse from(EchoDetail echoDetail) {
         return new MyMomentEchoResponse(echoDetail.id(), echoDetail.echoType(), echoDetail.userName());
